+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support `label` `image` attribute
 - Support `label` `justify` attribute (`:left`, `:center`, `:right`)
 - Support `label` `font` attribute (`:default`, `:text`, `:fixed`, `:menu`, `:heading`, `:caption`, `:small_caption`, `:icon`, `:tooltip`)
+- Support `label` `foreground` attribute
 
 ## 0.0.16