+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support `label` `justify` attribute (`:left`, `:center`, `:right`)
 - Support `label` `font` attribute (`:default`, `:text`, `:fixed`, `:menu`, `:heading`, `:caption`, `:small_caption`, `:icon`, `:tooltip`)
 - Support `label` `foreground` attribute
+- Support `label` `background` attribute (note that it does not work in Mac 'aqua' theme)
 
 ## 0.0.16