+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Support `label` `font` attribute (`:default`, `:text`, `:fixed`, `:menu`, `:heading`, `:caption`, `:small_caption`, `:icon`, `:tooltip`)
 - Support `label` `foreground` attribute
 - Support `label` `background` attribute (note that it does not work in Mac 'aqua' theme)
+- Support `label` `compound` attribute
+- Support `label` `relief` attributes (`'flat' (default), 'raised', 'sunken', 'solid', 'ridge', 'groove'`)
 
 ## 0.0.16